refactor: Move `.md`, `.rst` utils to `tools.markup.py`
Also some rearranging in `vega_expr.py`
