+++ tools/__init__.py
@@ -1,8 +1,15 @@
-from tools import generate_api_docs, generate_schema_wrapper, schemapi, update_init_file
+from tools import (
+    generate_api_docs,
+    generate_schema_wrapper,
+    markup,
+    schemapi,
+    update_init_file,
+)
 
 __all__ = [
     "generate_api_docs",
     "generate_schema_wrapper",
+    "markup",
     "schemapi",
     "update_init_file",
 ]
